Rearranged files, moved as much files to Utils.
