+++ src/PerformanceTests/Common/Scenarios/Receive/ReceiveRunner.V5.cs
@@ -1,7 +1,5 @@
-﻿#if Version6
+﻿#if Version5
 using NServiceBus;
-using System.Threading.Tasks;
-using Common.Scenarios;
 
 /// <summary>
 /// Does a continuous test where a pre-seeded amount of messages will be handled
@@ -9,17 +7,20 @@ using Common.Scenarios;
 partial class ReceiveRunner : ICreateSeedData
 {
     public int SeedSize { get; set; } = 50000;
-
+    
     public partial class Handler : IHandleMessages<Command>
     {
-        public async Task Handle(Command message, IMessageHandlerContext ctx)
+        public IBus Bus { get; set; }
+
+        public void Handle(Command message)
         {
         }
     }
 
-    public void SendMessage(IEndpointInstance endpointInstance, string endpointName)
+    public void SendMessage(ISendOnlyBus sendOnlyBus, string endpointName)
     {
-        endpointInstance.SendLocal(new Command());
+        var address = new Address(endpointName, "localhost");
+        sendOnlyBus.Send(address, new Command());
     }
 }
 #endif
