--- src/PerformanceTests/Common/TransportConfigOverride.cs
@@ -1,1 +1,0 @@
-﻿
