+++ src/PerformanceTests/Utils/EnumExtensionMethods.cs
@@ -11,7 +11,9 @@ namespace Tests
 
             var attributes = (DescriptionAttribute[])fi.GetCustomAttributes(typeof(DescriptionAttribute), false);
 
-            return attributes.Length > 0 ? attributes[0].Description : value.ToString();
+            return attributes.Length > 0
+                ? attributes[0].Description
+                : value.ToString();
         }
     }
 }
