+++ src/PerformanceTests/Utils/INeedPermutation.cs
