+++ src/PerformanceTests/Utils/Settings.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Configuration;
 
-static class Settings
+public static class Settings
 {
     public static TimeSpan WarmupDuration = TimeSpan.Parse(ConfigurationManager.AppSettings["WarmupDuration"]);
     public static TimeSpan RunDuration = TimeSpan.Parse(ConfigurationManager.AppSettings["RunDuration"]);
