+++ src/PerformanceTests/Utils/Utils.csproj
@@ -53,6 +53,7 @@
     <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.IO.Compression" />
+    <Reference Include="System.Messaging" />
     <Reference Include="System.Net.Http.WebRequest" />
     <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.ServiceModel" />
@@ -65,15 +66,21 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AssemblyScanner.cs" />
+    <Compile Include="ConcurrencyLevelConverter.cs" />
     <Compile Include="DeterministicUuid.cs" />
     <Compile Include="EnumExtensionMethods.cs" />
+    <Compile Include="INeedPermutation.cs" />
     <Compile Include="Permutations\Permutation.cs" />
     <Compile Include="Permutations\PermutationGenerator.cs" />
     <Compile Include="Permutations\PermutationParser.cs" />
     <Compile Include="Permutations\Permutations.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="QueueUtils.cs" />
+    <Compile Include="Settings.cs" />
     <Compile Include="Statistics.cs" />
     <Compile Include="SymbolicLink.cs" />
+    <Compile Include="TwoPhaseCommitEnlistment.cs" />
     <Compile Include="Variables\Audit.cs" />
     <Compile Include="Variables\ConcurrencyLevel.cs" />
     <Compile Include="Variables\DTC.cs" />
